requirements.txt
discord.py==2.5.2
python-dotenv==1.1.0
Requests==2.32.3
google-genai==1.16.1